Functioning check button now added that shows what setting QoS is currently set to

+++ gui.py
@@ -2,33 +2,29 @@ from dearpygui.core import *
 from dearpygui.simple import *
 import qos_settings
 
+settings_history = []
 
 def click_on(sender, data):
-    delete_item('text')
+    settings_history.append(1)
+    if is_item_shown('text') == True:
+        delete_item('text')
     qos_settings.click_radio_on()
     log_debug('qos on has been clicked...')
 
 
 def click_off(sender,data):
-    delete_item('text')
+    settings_history.append(0)
+    if is_item_shown('text') == True:
+        delete_item('text')
     qos_settings.click_radio_off()
     log_debug('qos off has been clicked...')
 
 
-# def check_current_settings(sender, data, settings_history=[]):
-#     try:
-#         delete_item('text1')
-#         delete_item('text')
-#     except:
-#         print('idk')
-#     if settings_history == []:
-#         settings_history.append(current)
-#         state = current
-#     elif settings_history[-1] == 0:
-#         state = 1
-#     elif settings_history[-1] == 1:
-#         state = 0
-#     add_text('text1', parent='Main', default_value=f'Current: {state}')
+def check_current_settings(sender, data):
+    if is_item_shown('text1') == True:
+        delete_item('text1')
+    else:
+        add_text('text1', parent='Main', default_value=f'QoS is now set to: {settings_history[-1]}')
 
 
 # GUI settings
@@ -42,7 +38,8 @@ set_style_window_padding(20,15)
 with window('Main'):
     print('Main window runnning...')
     add_text('Running')
-    # show_logger()
+    show_logger()
+    show_debug()
     qos_settings.init_driver()
     log_debug('Page loaded...')
     qos_settings.login_and_navigation()
@@ -57,7 +54,7 @@ with window('Main'):
     add_button('On', callback=click_on)
     add_button('Off', callback=click_off)
     add_spacing(count=4)
-    # add_button('Check', callback=check_current_settings)
+    add_button('Check', callback=check_current_settings)
     
 
 start_dearpygui(primary_window='Main')
